+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Last Changes
 
+- [#3](https://github.com/LaxarJS/laxar-infrastructure/issues/3): improve karma config
+
 
 ## v0.1.0-alpha.0